Handle zero-bytes in msgpack buffers. Fixes #20. NOT BACKWARDS COMPATIBLE
Turns out MessagePack messages can contain null-bytes
(that even makes a lot of sense, sorry for not noticing),
which means that using null-termination based string
conversion functions such as `packCString` in Haskell,
or `c_char_p` auto conversion in Python,
or `:string` auto conversion in Ruby,
cannot work (they will chop off the msgpack buffer
at the first null-byte).

This could trivially be observed by calling
an exported `Int -> Int` function with `0 :: Int`
(which serialises to '\NUL'), or using `1.0001 :: Double`,
whose serialisation also contains a null byte.

This commit fixes it by changing the calling convention:
Instead of sending just the packed `char *`, we prepend
it by a `int64_t` to indicate its length, so that
we can use e.g. `packCStringLen` on the Haskell side.

This makes the function wrappers on the calling side
slightly more complicated but the interface for using
the functions remains unchanged.

Also fix returning stack memory with
  `BS.useAsCString res_bs return`
That cannot work (unless we're lucky, apparently with
the exmaple Python and Ruby programs we were).
Instead return malloc'ed memory and free it at the
caller side.